:rocket: permissions on shop

+++ src/products/products.service.ts
@@ -2,6 +2,7 @@ import {
   BadRequestException,
   Injectable,
   NotFoundException,
+  UnauthorizedException,
 } from '@nestjs/common';
 import { PrismaService } from '../database/database.service';
 import { CreateProductDto } from './dto/create-product.dto';
@@ -11,13 +12,18 @@ import { UpdateProductDto } from './dto/update-product.dto';
 export class ProductsService {
   constructor(private readonly prisma: PrismaService) {}
 
-  async createProduct(createProductDto: CreateProductDto) {
+  async createProduct(createProductDto: CreateProductDto, req: any) {
+    const shop = await this.prisma.shop.findUnique({
+      where: { retailerId: req.id },
+    });
     const find = await this.prisma.product.findUnique({
       where: { name: createProductDto.name },
     });
     if (find) throw new BadRequestException('Ce nom est déjà pris');
 
-    return await this.prisma.product.create({ data: { ...createProductDto } });
+    return await this.prisma.product.create({
+      data: { ...createProductDto, shopId: shop.id },
+    });
   }
 
   async filterByPrice(min: number, max: number) {
@@ -47,9 +53,14 @@ export class ProductsService {
     });
   }
 
-  async deleteProduct(id: number) {
+  async deleteProduct(id: number, req: any) {
+    const shop = await this.prisma.shop.findUnique({
+      where: { retailerId: req.id },
+    });
     const product = await this.prisma.product.findUnique({ where: { id } });
     if (!product) throw new NotFoundException(`Ce produit n'existe pas`);
+
+    if (product.shopId != shop.id) throw new UnauthorizedException();
 
     return this.prisma.product.delete({ where: { id: product.id } });
   }
